move around various functionality into relevant modules to simplify elaborator module
