+++ compiler/noirc_frontend/src/elaborator/mod.rs
@@ -5,25 +5,15 @@ use std::{
 };
 
 use crate::{
-    DataType, NamedGeneric, StructField, TypeBindings,
-    ast::{IdentOrQuotedType, ItemVisibility, UnresolvedType},
-    graph::CrateGraph,
-    hir::def_collector::dc_crate::UnresolvedTrait,
-    hir_def::{
-        stmt::HirStatement,
-        traits::{NamedType, ResolvedTraitBound},
-    },
-    node_interner::{GlobalValue, QuotedTypeId},
-    token::SecondaryAttributeKind,
-    usage_tracker::UsageTracker,
+    NamedGeneric, ast::UnresolvedType, elaborator::globals::filter_literal_globals,
+    graph::CrateGraph, hir::def_collector::dc_crate::UnresolvedTrait, usage_tracker::UsageTracker,
     validity::length_is_zero,
 };
 use crate::{
-    EnumVariant, Shared, Type, TypeVariable,
+    Type, TypeVariable,
     ast::{
-        BlockExpression, FunctionKind, GenericTypeArgs, Ident, NoirFunction, NoirStruct, Param,
-        Path, Pattern, TraitBound, UnresolvedGeneric, UnresolvedGenerics,
-        UnresolvedTraitConstraint, UnresolvedTypeData, UnsupportedNumericGenericType, Visitor,
+        BlockExpression, FunctionKind, GenericTypeArgs, Ident, NoirFunction, Param,
+        UnresolvedGenerics, UnresolvedTypeData,
     },
     graph::CrateId,
     hir::{
@@ -31,26 +21,25 @@ use crate::{
         comptime::ComptimeError,
         def_collector::{
             dc_crate::{
-                CollectedItems, CompilationError, ImplMap, UnresolvedEnum, UnresolvedFunctions,
-                UnresolvedGlobal, UnresolvedStruct, UnresolvedTraitImpl, UnresolvedTypeAlias,
-                filter_literal_globals,
+                CollectedItems, CompilationError, ImplMap, UnresolvedFunctions, UnresolvedGlobal,
+                UnresolvedTraitImpl, UnresolvedTypeAlias,
             },
             errors::DefCollectorErrorKind,
         },
         def_map::{DefMaps, LocalModuleId, MAIN_FUNCTION, ModuleData, ModuleId},
         resolution::errors::ResolverError,
         scope::ScopeForest as GenericScopeForest,
-        type_check::{TypeCheckError, generics::TraitGenerics},
+        type_check::TypeCheckError,
     },
     hir_def::{
         expr::{HirCapturedVar, HirIdent},
         function::{FuncMeta, FunctionBody, HirFunction},
-        traits::{TraitConstraint, TraitImpl},
-        types::{Generics, Kind, ResolvedGeneric},
+        traits::TraitConstraint,
+        types::{Kind, ResolvedGeneric},
     },
     node_interner::{
-        DefinitionKind, DependencyId, FuncId, FunctionModifiers, GlobalId, NodeInterner,
-        ReferenceId, TraitId, TraitImplId, TypeAliasId, TypeId,
+        DefinitionKind, DependencyId, FuncId, FunctionModifiers, GlobalId, NodeInterner, TraitId,
+        TraitImplId, TypeAliasId, TypeId,
     },
     parser::{ParserError, ParserErrorReason},
 };
@@ -59,6 +48,9 @@ mod comptime;
 mod enums;
 mod expressions;
 mod function_context;
+mod generics;
+mod globals;
+mod impls;
 mod lints;
 mod options;
 mod path_resolution;
@@ -66,23 +58,22 @@ mod patterns;
 mod primitive_types;
 mod scope;
 mod statements;
+mod structs;
 mod trait_impls;
 mod traits;
 pub mod types;
 mod unquote;
+mod visibility;
 
 use function_context::FunctionContext;
-use im::HashSet;
 use iter_extended::vecmap;
-use noirc_errors::{Located, Location};
+use noirc_errors::Location;
 pub(crate) use options::ElaboratorOptions;
 pub use options::{FrontendOptions, UnstableFeature};
 pub use path_resolution::Turbofish;
 use path_resolution::{
     PathResolution, PathResolutionItem, PathResolutionMode, PathResolutionTarget,
 };
-use rustc_hash::FxHashMap as HashMap;
-use types::bind_ordered_generics;
 
 use self::traits::check_trait_impl_method_matches_declaration;
 pub(crate) use path_resolution::{TypedPath, TypedPathSegment};
@@ -406,15 +397,6 @@ impl<'context> Elaborator<'context> {
         self.options.pedantic_solving
     }
 
-    /// Runs `f` and if it modifies `self.generics`, `self.generics` is truncated
-    /// back to the previous length.
-    fn recover_generics<T>(&mut self, f: impl FnOnce(&mut Self) -> T) -> T {
-        let generics_count = self.generics.len();
-        let ret = f(self);
-        self.generics.truncate(generics_count);
-        ret
-    }
-
     fn elaborate_functions(&mut self, functions: UnresolvedFunctions) {
         for (_, id, _) in functions.functions {
             self.elaborate_function(id);
@@ -422,26 +404,6 @@ impl<'context> Elaborator<'context> {
 
         self.generics.clear();
         self.self_type = None;
-    }
-
-    fn introduce_generics_into_scope(&mut self, all_generics: Vec<ResolvedGeneric>) {
-        // Introduce all numeric generics into scope
-        for generic in &all_generics {
-            if let Kind::Numeric(typ) = &generic.kind() {
-                let definition =
-                    DefinitionKind::NumericGeneric(generic.type_var.clone(), typ.clone());
-                let ident = Ident::new(generic.name.to_string(), generic.location);
-                let hir_ident = self.add_variable_decl(
-                    ident, false, // mutable
-                    false, // allow_shadowing
-                    false, // warn_if_unused
-                    definition,
-                );
-                self.interner.push_definition_type(hir_ident.id, *typ.clone());
-            }
-        }
-
-        self.generics = all_generics;
     }
 
     pub(crate) fn elaborate_function(&mut self, id: FuncId) {
@@ -590,139 +552,6 @@ impl<'context> Elaborator<'context> {
         self.current_item = old_item;
     }
 
-    /// This turns function parameters of the form:
-    /// `fn foo(x: impl Bar)`
-    ///
-    /// into
-    /// `fn foo<T0_impl_Bar>(x: T0_impl_Bar) where T0_impl_Bar: Bar`
-    /// although the fresh type variable is not named internally.
-    fn desugar_impl_trait_arg(
-        &mut self,
-        trait_path: Path,
-        trait_generics: GenericTypeArgs,
-        generics: &mut Vec<TypeVariable>,
-        trait_constraints: &mut Vec<TraitConstraint>,
-    ) -> Type {
-        let new_generic_id = self.interner.next_type_variable_id();
-
-        let new_generic = TypeVariable::unbound(new_generic_id, Kind::Normal);
-        generics.push(new_generic.clone());
-
-        let name = format!("impl {trait_path}");
-        let generic_type = Type::NamedGeneric(NamedGeneric {
-            type_var: new_generic,
-            name: Rc::new(name),
-            implicit: false,
-        });
-        let trait_bound = TraitBound { trait_path, trait_id: None, trait_generics };
-
-        if let Some(trait_bound) = self.resolve_trait_bound(&trait_bound) {
-            let new_constraint = TraitConstraint { typ: generic_type.clone(), trait_bound };
-            trait_constraints.push(new_constraint);
-        }
-
-        generic_type
-    }
-
-    /// Add the given generics to scope.
-    /// Each generic will have a fresh `Shared<TypeBinding>` associated with it.
-    pub fn add_generics(&mut self, generics: &UnresolvedGenerics) -> Generics {
-        vecmap(generics, |generic| {
-            let mut is_error = false;
-            let (type_var, name) = match self.resolve_generic(generic) {
-                Ok(values) => values,
-                Err(error) => {
-                    self.push_err(error);
-                    is_error = true;
-                    let id = self.interner.next_type_variable_id();
-                    let kind = self.resolve_generic_kind(generic);
-                    (TypeVariable::unbound(id, kind), Rc::new("(error)".into()))
-                }
-            };
-
-            let location = generic.location();
-            let name_owned = name.as_ref().clone();
-            let resolved_generic = ResolvedGeneric { name, type_var, location };
-
-            // Check for name collisions of this generic
-            // Checking `is_error` here prevents DuplicateDefinition errors when
-            // we have multiple generics from macros which fail to resolve and
-            // are all given the same default name "(error)".
-            if !is_error {
-                if let Some(generic) = self.find_generic(&name_owned) {
-                    self.push_err(ResolverError::DuplicateDefinition {
-                        name: name_owned,
-                        first_location: generic.location,
-                        second_location: location,
-                    });
-                } else {
-                    self.generics.push(resolved_generic.clone());
-                }
-            }
-
-            resolved_generic
-        })
-    }
-
-    fn resolve_generic(
-        &mut self,
-        generic: &UnresolvedGeneric,
-    ) -> Result<(TypeVariable, Rc<String>), ResolverError> {
-        // Map the generic to a fresh type variable
-        match generic.ident() {
-            IdentOrQuotedType::Ident(ident) => {
-                let id = self.interner.next_type_variable_id();
-                let kind = self.resolve_generic_kind(generic);
-                let typevar = TypeVariable::unbound(id, kind);
-                let name = Rc::new(ident.to_string());
-                Ok((typevar, name))
-            }
-            IdentOrQuotedType::Quoted(id, location) => {
-                match self.interner.get_quoted_type(*id).follow_bindings() {
-                    Type::NamedGeneric(NamedGeneric { type_var, name, .. }) => {
-                        Ok((type_var.clone(), name))
-                    }
-                    other => Err(ResolverError::MacroResultInGenericsListNotAGeneric {
-                        location: *location,
-                        typ: other.clone(),
-                    }),
-                }
-            }
-        }
-    }
-
-    /// Return the kind of an unresolved generic.
-    /// If a numeric generic has been specified, resolve the annotated type to make
-    /// sure only primitive numeric types are being used.
-    pub(super) fn resolve_generic_kind(&mut self, generic: &UnresolvedGeneric) -> Kind {
-        if let UnresolvedGeneric::Numeric { ident, typ } = generic {
-            let unresolved_typ = typ.clone();
-            let wildcard_allowed = false;
-            let typ = if unresolved_typ.is_type_expression() {
-                self.resolve_type_with_kind(
-                    unresolved_typ.clone(),
-                    &Kind::numeric(Type::default_int_type()),
-                    wildcard_allowed,
-                )
-            } else {
-                self.resolve_type(unresolved_typ.clone(), wildcard_allowed)
-            };
-            if !matches!(typ, Type::FieldElement | Type::Integer(_, _)) {
-                let unsupported_typ_err =
-                    ResolverError::UnsupportedNumericGenericType(UnsupportedNumericGenericType {
-                        name: ident.ident().map(|name| name.to_string()),
-                        typ: typ.to_string(),
-                        location: unresolved_typ.location,
-                    });
-
-                self.push_err(unsupported_typ_err);
-            }
-            Kind::numeric(typ)
-        } else {
-            Kind::Normal
-        }
-    }
-
     pub(crate) fn push_err(&mut self, error: impl Into<CompilationError>) {
         let error: CompilationError = error.into();
         self.errors.push(error);
@@ -762,162 +591,6 @@ impl<'context> Elaborator<'context> {
         };
         self.push_err(error);
         None
-    }
-
-    /// Resolve the given trait constraints and add them to scope as we go.
-    /// This second step is necessary to resolve subsequent constraints such
-    /// as `<T as Foo>::Bar: Eq` which may lookup an impl which was assumed
-    /// by a previous constraint.
-    ///
-    /// If these constraints are unwanted afterward they should be manually
-    /// removed from the interner.
-    fn resolve_trait_constraints(
-        &mut self,
-        where_clause: &[UnresolvedTraitConstraint],
-    ) -> Vec<TraitConstraint> {
-        where_clause
-            .iter()
-            .filter_map(|constraint| self.resolve_trait_constraint(constraint))
-            .collect()
-    }
-
-    /// Expands any traits in a where clause to mention all associated types if they were
-    /// elided by the user. See `add_missing_named_generics` for more  detail.
-    ///
-    /// Returns all newly created generics to be added to this function/trait/impl.
-    fn desugar_trait_constraints(
-        &mut self,
-        where_clause: &mut [UnresolvedTraitConstraint],
-    ) -> Vec<(ResolvedGeneric, Vec<ResolvedTraitBound>)> {
-        where_clause
-            .iter_mut()
-            .flat_map(|constraint| {
-                self.add_missing_named_generics(&constraint.typ, &mut constraint.trait_bound)
-            })
-            .collect()
-    }
-
-    /// For each associated type that isn't mentioned in a trait bound, this adds
-    /// the type as an implicit generic to the where clause and returns the newly
-    /// created generics in a vector to add to the function/trait/impl later.
-    /// For example, this will turn a function using a trait with 2 associated types:
-    ///
-    /// `fn foo<T>() where T: Foo { ... }`
-    ///
-    /// into:
-    /// `fn foo<T>() where T: Foo<Bar = A, Baz = B> { ... }`
-    ///
-    /// with a vector of `<A, B>` returned so that the caller can then modify the function to:
-    /// `fn foo<T, A, B>() where T: Foo<Bar = A, Baz = B> { ... }`
-    fn add_missing_named_generics(
-        &mut self,
-        object: &UnresolvedType,
-        bound: &mut TraitBound,
-    ) -> Vec<(ResolvedGeneric, Vec<ResolvedTraitBound>)> {
-        let mut added_generics = Vec::new();
-        let trait_path = self.validate_path(bound.trait_path.clone());
-
-        let Ok(PathResolutionItem::Trait(trait_id)) =
-            self.resolve_path_or_error(trait_path, PathResolutionTarget::Type)
-        else {
-            return Vec::new();
-        };
-
-        let the_trait = self.get_trait_mut(trait_id);
-
-        if the_trait.associated_types.len() > bound.trait_generics.named_args.len() {
-            let trait_name = the_trait.name.to_string();
-            let associated_type_bounds = the_trait.associated_type_bounds.clone();
-
-            for associated_type in &the_trait.associated_types.clone() {
-                if !bound
-                    .trait_generics
-                    .named_args
-                    .iter()
-                    .any(|(name, _)| name.as_str() == *associated_type.name.as_ref())
-                {
-                    // This generic isn't contained in the bound's named arguments,
-                    // so add it by creating a fresh type variable.
-                    let new_generic_id = self.interner.next_type_variable_id();
-                    let kind = associated_type.type_var.kind();
-                    let type_var = TypeVariable::unbound(new_generic_id, kind);
-
-                    let location = bound.trait_path.location;
-                    let name = format!("<{object} as {trait_name}>::{}", associated_type.name);
-                    let name = Rc::new(name);
-                    let typ = Type::NamedGeneric(NamedGeneric {
-                        type_var: type_var.clone(),
-                        name: name.clone(),
-                        implicit: true,
-                    });
-                    let typ = self.interner.push_quoted_type(typ);
-                    let typ = UnresolvedTypeData::Resolved(typ).with_location(location);
-                    let ident = Ident::new(associated_type.name.as_ref().clone(), location);
-
-                    let associated_type_bounds = associated_type_bounds
-                        .get(associated_type.name.as_str())
-                        .cloned()
-                        .unwrap_or_default();
-
-                    bound.trait_generics.named_args.push((ident, typ));
-                    added_generics.push((
-                        ResolvedGeneric { name, location, type_var },
-                        associated_type_bounds,
-                    ));
-                }
-            }
-        }
-
-        added_generics
-    }
-
-    /// Resolves a trait constraint and adds it to scope as an assumed impl.
-    /// This second step is necessary to resolve subsequent constraints such
-    /// as `<T as Foo>::Bar: Eq` which may lookup an impl which was assumed
-    /// by a previous constraint.
-    fn resolve_trait_constraint(
-        &mut self,
-        constraint: &UnresolvedTraitConstraint,
-    ) -> Option<TraitConstraint> {
-        let wildcard_allowed = false;
-        let typ = self.resolve_type(constraint.typ.clone(), wildcard_allowed);
-        let trait_bound = self.resolve_trait_bound(&constraint.trait_bound)?;
-        let location = constraint.trait_bound.trait_path.location;
-
-        self.add_trait_bound_to_scope(location, &typ, &trait_bound, trait_bound.trait_id);
-
-        Some(TraitConstraint { typ, trait_bound })
-    }
-
-    pub fn resolve_trait_bound(&mut self, bound: &TraitBound) -> Option<ResolvedTraitBound> {
-        self.resolve_trait_bound_inner(bound, PathResolutionMode::MarkAsReferenced)
-    }
-
-    pub fn use_trait_bound(&mut self, bound: &TraitBound) -> Option<ResolvedTraitBound> {
-        self.resolve_trait_bound_inner(bound, PathResolutionMode::MarkAsUsed)
-    }
-
-    fn resolve_trait_bound_inner(
-        &mut self,
-        bound: &TraitBound,
-        mode: PathResolutionMode,
-    ) -> Option<ResolvedTraitBound> {
-        let trait_path = self.validate_path(bound.trait_path.clone());
-        let the_trait = self.lookup_trait_or_error(trait_path)?;
-        let trait_id = the_trait.id;
-        let location = bound.trait_path.location;
-        let wildcard_allowed = false;
-
-        let (ordered, named) = self.resolve_type_args_inner(
-            bound.trait_generics.clone(),
-            trait_id,
-            location,
-            mode,
-            wildcard_allowed,
-        );
-
-        let trait_generics = TraitGenerics { ordered, named };
-        Some(ResolvedTraitBound { trait_id, trait_generics, location })
     }
 
     /// Extract metadata from a NoirFunction
@@ -1261,94 +934,6 @@ impl<'context> Elaborator<'context> {
         }
     }
 
-    fn add_trait_constraints_to_scope<'a>(
-        &mut self,
-        constraints: impl Iterator<Item = &'a TraitConstraint>,
-        location: Location,
-    ) {
-        for constraint in constraints {
-            self.add_trait_bound_to_scope(
-                location,
-                &constraint.typ,
-                &constraint.trait_bound,
-                constraint.trait_bound.trait_id,
-            );
-        }
-
-        // Also assume `self` implements the current trait if we are inside a trait definition
-        if let Some(trait_id) = self.current_trait {
-            let the_trait = self.interner.get_trait(trait_id);
-            let constraint = the_trait.as_constraint(the_trait.name.location());
-            let self_type =
-                self.self_type.clone().expect("Expected a self type if there's a current trait");
-            self.add_trait_bound_to_scope(
-                location,
-                &self_type,
-                &constraint.trait_bound,
-                constraint.trait_bound.trait_id,
-            );
-        }
-    }
-
-    fn remove_trait_constraints_from_scope<'a>(
-        &mut self,
-        constraints: impl Iterator<Item = &'a TraitConstraint>,
-    ) {
-        for constraint in constraints {
-            self.interner
-                .remove_assumed_trait_implementations_for_trait(constraint.trait_bound.trait_id);
-        }
-
-        // Also remove the assumed trait implementation for `self` if this is a trait definition
-        if let Some(trait_id) = self.current_trait {
-            self.interner.remove_assumed_trait_implementations_for_trait(trait_id);
-        }
-    }
-
-    fn add_trait_bound_to_scope(
-        &mut self,
-        location: Location,
-        object: &Type,
-        trait_bound: &ResolvedTraitBound,
-        starting_trait_id: TraitId,
-    ) {
-        let trait_id = trait_bound.trait_id;
-        let generics = trait_bound.trait_generics.clone();
-
-        if !self.interner.add_assumed_trait_implementation(object.clone(), trait_id, generics) {
-            if let Some(the_trait) = self.interner.try_get_trait(trait_id) {
-                let trait_name = the_trait.name.to_string();
-                let typ = object.clone();
-                self.push_err(TypeCheckError::UnneededTraitConstraint {
-                    trait_name,
-                    typ,
-                    location,
-                });
-            }
-        }
-
-        // Also add assumed implementations for the parent traits, if any
-        if let Some(trait_bounds) =
-            self.interner.try_get_trait(trait_id).map(|the_trait| the_trait.trait_bounds.clone())
-        {
-            for parent_trait_bound in trait_bounds {
-                // Avoid looping forever in case there are cycles
-                if parent_trait_bound.trait_id == starting_trait_id {
-                    continue;
-                }
-
-                let parent_trait_bound =
-                    self.instantiate_parent_trait_bound(trait_bound, &parent_trait_bound);
-                self.add_trait_bound_to_scope(
-                    location,
-                    object,
-                    &parent_trait_bound,
-                    starting_trait_id,
-                );
-            }
-        }
-    }
-
     fn elaborate_traits(&mut self, traits: BTreeMap<TraitId, UnresolvedTrait>) {
         for (trait_id, unresolved_trait) in traits {
             self.current_trait = Some(trait_id);
@@ -1391,394 +976,6 @@ impl<'context> Elaborator<'context> {
         self.generics.clear();
     }
 
-    fn add_trait_impl_assumed_trait_implementations(&mut self, impl_id: Option<TraitImplId>) {
-        if let Some(impl_id) = impl_id {
-            if let Some(trait_implementation) = self.interner.try_get_trait_implementation(impl_id)
-            {
-                for trait_constrain in &trait_implementation.borrow().where_clause {
-                    let trait_bound = &trait_constrain.trait_bound;
-                    self.add_trait_bound_to_scope(
-                        trait_bound.location,
-                        &trait_constrain.typ,
-                        trait_bound,
-                        trait_bound.trait_id,
-                    );
-                }
-            }
-        }
-    }
-
-    fn remove_trait_impl_assumed_trait_implementations(&mut self, impl_id: Option<TraitImplId>) {
-        if let Some(impl_id) = impl_id {
-            if let Some(trait_implementation) = self.interner.try_get_trait_implementation(impl_id)
-            {
-                for trait_constrain in &trait_implementation.borrow().where_clause {
-                    self.interner.remove_assumed_trait_implementations_for_trait(
-                        trait_constrain.trait_bound.trait_id,
-                    );
-                }
-            }
-        }
-    }
-
-    fn check_trait_impl_where_clause_matches_trait_where_clause(
-        &mut self,
-        trait_impl: &UnresolvedTraitImpl,
-    ) {
-        let Some(trait_id) = trait_impl.trait_id else {
-            return;
-        };
-
-        let Some(the_trait) = self.interner.try_get_trait(trait_id) else {
-            return;
-        };
-
-        if the_trait.where_clause.is_empty() {
-            return;
-        }
-
-        let impl_trait = the_trait.name.to_string();
-
-        let mut bindings = TypeBindings::default();
-        bind_ordered_generics(
-            &the_trait.generics,
-            &trait_impl.resolved_trait_generics,
-            &mut bindings,
-        );
-
-        // Check that each of the trait's where clause constraints is satisfied
-        for trait_constraint in the_trait.where_clause.clone() {
-            let Some(trait_constraint_trait) =
-                self.interner.try_get_trait(trait_constraint.trait_bound.trait_id)
-            else {
-                continue;
-            };
-            let trait_constraint_trait_name = trait_constraint_trait.name.to_string();
-
-            let mut trait_constraint = trait_constraint.clone();
-            trait_constraint.apply_bindings(&bindings);
-
-            let trait_constraint_type = trait_constraint.typ;
-            let trait_bound = trait_constraint.trait_bound;
-
-            let mut named_generics = trait_bound.trait_generics.named.clone();
-
-            // If the trait bound is over a trait that has associated types, the ones that
-            // aren't explicit will be in `named_generics` as implicitly added ones.
-            // If they are unbound, they won't be bound until monomorphization, in which cae
-            // the below trait implementation lookup will fail (an unbound named generic will
-            // never unify in this case). In this case we replace them with fresh type variables
-            // so they'll unify (the bindings aren't applied here so this is fine).
-            // If they are bound though, we won't replace them as we want to ensure the binding
-            // matches.
-            //
-            // `bindings` is passed here because these implicitly added named generics might
-            // have a constraint on them later on and we want to remember what type they ended
-            // up being.
-            self.replace_implicitly_added_unbound_named_generics_with_fresh_type_variables(
-                &mut named_generics,
-                &mut bindings,
-            );
-
-            match self.interner.try_lookup_trait_implementation(
-                &trait_constraint_type,
-                trait_bound.trait_id,
-                &trait_bound.trait_generics.ordered,
-                &named_generics,
-            ) {
-                Ok((_, impl_bindings, impl_instantiation_bindings)) => {
-                    bindings.extend(impl_bindings);
-                    bindings.extend(impl_instantiation_bindings);
-                }
-                Err(_) => {
-                    let missing_trait =
-                        format!("{}{}", trait_constraint_trait_name, trait_bound.trait_generics);
-                    self.push_err(ResolverError::TraitNotImplemented {
-                        impl_trait: impl_trait.clone(),
-                        missing_trait,
-                        type_missing_trait: trait_constraint_type.to_string(),
-                        location: trait_impl.object_type.location,
-                        missing_trait_location: trait_bound.location,
-                    });
-                }
-            }
-        }
-    }
-
-    fn replace_implicitly_added_unbound_named_generics_with_fresh_type_variables(
-        &mut self,
-        named_generics: &mut [NamedType],
-        bindings: &mut TypeBindings,
-    ) {
-        for named_type in named_generics.iter_mut() {
-            match &named_type.typ {
-                Type::NamedGeneric(NamedGeneric { type_var, implicit: true, .. })
-                    if type_var.borrow().is_unbound() =>
-                {
-                    let type_var_id = type_var.id();
-                    let new_type_var_id = self.interner.next_type_variable_id();
-                    let kind = type_var.kind();
-                    let new_type_var = TypeVariable::unbound(new_type_var_id, kind.clone());
-                    named_type.typ = Type::TypeVariable(new_type_var.clone());
-                    bindings.insert(type_var_id, (new_type_var, kind, named_type.typ.clone()));
-                }
-                _ => (),
-            };
-        }
-    }
-
-    fn check_parent_traits_are_implemented(&mut self, trait_impl: &UnresolvedTraitImpl) {
-        let Some(trait_id) = trait_impl.trait_id else {
-            return;
-        };
-
-        let Some(object_type) = &trait_impl.resolved_object_type else {
-            return;
-        };
-
-        let Some(the_trait) = self.interner.try_get_trait(trait_id) else {
-            return;
-        };
-
-        if the_trait.trait_bounds.is_empty() {
-            return;
-        }
-
-        let impl_trait = the_trait.name.to_string();
-
-        let mut bindings = TypeBindings::default();
-        bind_ordered_generics(
-            &the_trait.generics,
-            &trait_impl.resolved_trait_generics,
-            &mut bindings,
-        );
-
-        // Note: we only check if the immediate parents are implemented, we don't check recursively.
-        // Why? If a parent isn't implemented, we get an error. If a parent is implemented, we'll
-        // do the same check for the parent, so this trait's parents parents will be checked, so the
-        // recursion is guaranteed.
-        for parent_trait_bound in the_trait.trait_bounds.clone() {
-            let Some(parent_trait) = self.interner.try_get_trait(parent_trait_bound.trait_id)
-            else {
-                continue;
-            };
-
-            let parent_trait_bound = ResolvedTraitBound {
-                trait_generics: parent_trait_bound
-                    .trait_generics
-                    .map(|typ| typ.substitute(&bindings)),
-                ..parent_trait_bound
-            };
-
-            if self
-                .interner
-                .try_lookup_trait_implementation(
-                    object_type,
-                    parent_trait_bound.trait_id,
-                    &parent_trait_bound.trait_generics.ordered,
-                    &parent_trait_bound.trait_generics.named,
-                )
-                .is_err()
-            {
-                let missing_trait =
-                    format!("{}{}", parent_trait.name, parent_trait_bound.trait_generics);
-                self.push_err(ResolverError::TraitNotImplemented {
-                    impl_trait: impl_trait.clone(),
-                    missing_trait,
-                    type_missing_trait: trait_impl.object_type.to_string(),
-                    location: trait_impl.object_type.location,
-                    missing_trait_location: parent_trait_bound.location,
-                });
-            }
-        }
-    }
-
-    fn collect_impls(
-        &mut self,
-        module: LocalModuleId,
-        impls: &mut [(UnresolvedGenerics, Location, UnresolvedFunctions)],
-        self_type: &UnresolvedType,
-    ) {
-        self.local_module = module;
-
-        for (generics, location, unresolved) in impls {
-            self.check_generics_appear_in_types(generics, &[self_type], &[]);
-
-            let old_generic_count = self.generics.len();
-            self.add_generics(generics);
-            self.declare_methods_on_struct(None, unresolved, *location);
-            self.generics.truncate(old_generic_count);
-        }
-    }
-
-    fn collect_trait_impl(&mut self, trait_impl: &mut UnresolvedTraitImpl) {
-        self.local_module = trait_impl.module_id;
-        self.current_trait_impl = trait_impl.impl_id;
-
-        let self_type = trait_impl.methods.self_type.clone();
-        let self_type =
-            self_type.expect("Expected struct type to be set before collect_trait_impl");
-
-        self.self_type = Some(self_type.clone());
-        let self_type_location = trait_impl.object_type.location;
-
-        if matches!(self_type, Type::Reference(..)) {
-            self.push_err(DefCollectorErrorKind::ReferenceInTraitImpl {
-                location: self_type_location,
-            });
-        }
-
-        self.check_generics_appear_in_types(
-            &trait_impl.generics,
-            &[&trait_impl.r#trait, &trait_impl.object_type],
-            &trait_impl.where_clause,
-        );
-
-        if let Some(trait_id) = trait_impl.trait_id {
-            self.generics = trait_impl.resolved_generics.clone();
-
-            let where_clause = self.resolve_trait_constraints(&trait_impl.where_clause);
-
-            // Now solve the actual type of associated types
-            // (before this we only declared them without knowing their type)
-            if let Some(trait_impl_id) = trait_impl.impl_id {
-                let unresolved_associated_types =
-                    std::mem::take(&mut trait_impl.unresolved_associated_types);
-                let mut unresolved_associated_types =
-                    unresolved_associated_types.into_iter().collect::<HashMap<_, _>>();
-
-                let associated_types =
-                    self.interner.get_associated_types_for_impl(trait_impl_id).to_vec();
-                for associated_type in &associated_types {
-                    let Type::NamedGeneric(named_generic) = &associated_type.typ else {
-                        // This can happen if the associated type is specified directly in the impl trait generics,
-                        // This can't be done in code, but it could happen with unquoted types.
-                        continue;
-                    };
-
-                    let Some(unresolved_type) =
-                        unresolved_associated_types.remove(&associated_type.name)
-                    else {
-                        // This too can happen if the associated type is specified directly in the impl trait generics,
-                        // like `impl<H> BuildHasher<H = H>`, where `H` is a named generic but its resolution isn't delayed.
-                        // This can't be done in code, but it could happen with unquoted types.
-                        continue;
-                    };
-                    let wildcard_allowed = false;
-                    let resolved_type = self.resolve_type_with_kind(
-                        unresolved_type,
-                        &associated_type.typ.kind(),
-                        wildcard_allowed,
-                    );
-                    named_generic.type_var.bind(resolved_type);
-                }
-            }
-
-            let trait_ = self.interner.get_trait(trait_id);
-
-            // If there are bounds on the trait's associated types, check them now
-            let associated_type_bounds = &trait_.associated_type_bounds;
-            if !associated_type_bounds.is_empty() {
-                let associated_type_bounds = associated_type_bounds.clone();
-                let named_generics = self
-                    .interner
-                    .get_associated_types_for_impl(trait_impl.impl_id.unwrap())
-                    .to_vec();
-                for named_generic in named_generics {
-                    let Some(bounds) = associated_type_bounds.get(named_generic.name.as_str())
-                    else {
-                        continue;
-                    };
-                    let object_type = &named_generic.typ;
-                    for bound in bounds {
-                        if let Err(error) = self.interner.lookup_trait_implementation(
-                            object_type,
-                            bound.trait_id,
-                            &bound.trait_generics.ordered,
-                            &bound.trait_generics.named,
-                        ) {
-                            self.push_trait_constraint_error(
-                                object_type,
-                                error,
-                                named_generic.name.location(),
-                            );
-                        }
-                    }
-                }
-            }
-
-            self.remove_trait_constraints_from_scope(where_clause.iter());
-
-            self.collect_trait_impl_methods(trait_id, trait_impl, &where_clause);
-
-            let location = trait_impl.object_type.location;
-            self.declare_methods_on_struct(Some(trait_id), &mut trait_impl.methods, location);
-
-            let trait_visibility = self.interner.get_trait(trait_id).visibility;
-
-            let methods = trait_impl.methods.function_ids();
-            for func_id in &methods {
-                self.interner.set_function_trait(*func_id, self_type.clone(), trait_id);
-
-                // A trait impl method has the same visibility as its trait
-                let modifiers = self.interner.function_modifiers_mut(func_id);
-                modifiers.visibility = trait_visibility;
-            }
-
-            let trait_generics = trait_impl.resolved_trait_generics.clone();
-            let ident = match &trait_impl.r#trait.typ {
-                UnresolvedTypeData::Named(trait_path, _, _) => trait_path.last_ident(),
-                UnresolvedTypeData::Resolved(quoted_type_id) => {
-                    let typ = self.interner.get_quoted_type(*quoted_type_id);
-                    let name = if let Type::TraitAsType(_, name, _) = typ {
-                        name.to_string()
-                    } else {
-                        typ.to_string()
-                    };
-                    Ident::new(name, trait_impl.r#trait.location)
-                }
-                _ => {
-                    // We don't error in this case because an error will be produced later on when
-                    // solving the trait impl trait type
-                    Ident::new(trait_impl.r#trait.to_string(), trait_impl.r#trait.location)
-                }
-            };
-
-            let resolved_trait_impl = Shared::new(TraitImpl {
-                ident,
-                location,
-                typ: self_type.clone(),
-                trait_id,
-                trait_generics,
-                file: trait_impl.file_id,
-                crate_id: self.crate_id,
-                where_clause,
-                methods,
-            });
-
-            let generics = vecmap(&self.generics, |generic| generic.type_var.clone());
-
-            if let Err(prev_location) = self.interner.add_trait_implementation(
-                self_type.clone(),
-                trait_id,
-                trait_impl.impl_id.expect("impl_id should be set in define_function_metas"),
-                generics,
-                resolved_trait_impl,
-            ) {
-                self.push_err(DefCollectorErrorKind::OverlappingImpl {
-                    typ: self_type.clone(),
-                    location: self_type_location,
-                    prev_location,
-                });
-            }
-        }
-
-        self.generics.clear();
-
-        self.current_trait_impl = None;
-        self.self_type = None;
-    }
-
     pub fn get_module(&self, module: ModuleId) -> &ModuleData {
         let message = "A crate should always be present for a given crate id";
         &self.def_maps.get(&module.krate).expect(message)[module.local_id]
@@ -1787,94 +984,6 @@ impl<'context> Elaborator<'context> {
     fn get_module_mut(def_maps: &mut DefMaps, module: ModuleId) -> &mut ModuleData {
         let message = "A crate should always be present for a given crate id";
         &mut def_maps.get_mut(&module.krate).expect(message)[module.local_id]
-    }
-
-    fn declare_methods_on_struct(
-        &mut self,
-        trait_id: Option<TraitId>,
-        functions: &mut UnresolvedFunctions,
-        location: Location,
-    ) {
-        let self_type = functions.self_type.as_ref();
-        let self_type =
-            self_type.expect("Expected struct type to be set before declare_methods_on_struct");
-
-        let function_ids = functions.function_ids();
-
-        if let Type::DataType(struct_type, _) = &self_type {
-            let struct_ref = struct_type.borrow();
-
-            // `impl`s are only allowed on types defined within the current crate
-            if trait_id.is_none() && struct_ref.id.krate() != self.crate_id {
-                let type_name = struct_ref.name.to_string();
-                self.push_err(DefCollectorErrorKind::ForeignImpl { location, type_name });
-                return;
-            }
-
-            // Grab the module defined by the struct type. Note that impls are a case
-            // where the module the methods are added to is not the same as the module
-            // they are resolved in.
-            let module = Self::get_module_mut(self.def_maps, struct_ref.id.module_id());
-
-            for (_, method_id, method) in &functions.functions {
-                // If this method was already declared, remove it from the module so it cannot
-                // be accessed with the `TypeName::method` syntax. We'll check later whether the
-                // object types in each method overlap or not. If they do, we issue an error.
-                // If not, that is specialization which is allowed.
-                let name = method.name_ident().clone();
-                let result = if let Some(trait_id) = trait_id {
-                    module.declare_trait_function(name, *method_id, trait_id)
-                } else {
-                    module.declare_function(name, method.def.visibility, *method_id)
-                };
-                if result.is_err() {
-                    let existing = module.find_func_with_name(method.name_ident()).expect(
-                        "declare_function should only error if there is an existing function",
-                    );
-
-                    // Only remove the existing function from scope if it is from a trait impl as
-                    // well. If it is from a non-trait impl that should override trait impl methods
-                    // anyway so that Foo::bar always resolves to the non-trait impl version.
-                    if self.interner.function_meta(&existing).trait_impl.is_some() {
-                        module.remove_function(method.name_ident());
-                    }
-                }
-            }
-
-            // Trait impl methods are already declared in NodeInterner::add_trait_implementation
-            if trait_id.is_none() {
-                self.declare_methods(self_type, &function_ids);
-            }
-        // We can define methods on primitive types only if we're in the stdlib
-        } else if trait_id.is_none() && *self_type != Type::Error {
-            if self.crate_id.is_stdlib() {
-                // Trait impl methods are already declared in NodeInterner::add_trait_implementation
-                if trait_id.is_none() {
-                    self.declare_methods(self_type, &function_ids);
-                }
-            } else {
-                self.push_err(DefCollectorErrorKind::NonStructTypeInImpl { location });
-            }
-        }
-    }
-
-    fn declare_methods(&mut self, self_type: &Type, function_ids: &[FuncId]) {
-        for method_id in function_ids {
-            let method_name = self.interner.function_name(method_id).to_owned();
-
-            if let Some(first_fn) =
-                self.interner.add_method(self_type, method_name.clone(), *method_id, None)
-            {
-                let first_location = self.interner.function_ident(&first_fn).location();
-                let second_location = self.interner.function_ident(method_id).location();
-                let error = ResolverError::DuplicateDefinition {
-                    name: method_name,
-                    first_location,
-                    second_location,
-                };
-                self.push_err(error);
-            }
-        }
     }
 
     fn define_type_alias(&mut self, alias_id: TypeAliasId, alias: UnresolvedTypeAlias) {
@@ -1924,371 +1033,11 @@ impl<'context> Elaborator<'context> {
             (self.use_type(alias.type_alias_def.typ, wildcard_allowed), None)
         };
 
-        if visibility != ItemVisibility::Private {
+        if !visibility.is_private() {
             self.check_type_is_not_more_private_then_item(name, visibility, &typ, location);
         }
         self.interner.set_type_alias(alias_id, typ, generics, num_expr);
         self.generics.clear();
-    }
-
-    /// Find the struct in the parent module so we can know its visibility
-    fn find_struct_visibility(&self, struct_type: &DataType) -> Option<ItemVisibility> {
-        let parent_module_id = struct_type.id.parent_module_id(self.def_maps);
-        let parent_module_data = self.get_module(parent_module_id);
-        let per_ns = parent_module_data.find_name(&struct_type.name);
-        per_ns.types.map(|(_, vis, _)| vis)
-    }
-
-    /// Check whether a functions return value and args should be checked for private type visibility.
-    fn should_check_function_visibility(
-        &self,
-        func_meta: &FuncMeta,
-        modifiers: &FunctionModifiers,
-    ) -> bool {
-        // Private functions don't leak anything.
-        if modifiers.visibility == ItemVisibility::Private {
-            return false;
-        }
-        // Implementing public traits on private types is okay, they can't be used unless the type itself is accessible.
-        if func_meta.trait_impl.is_some() {
-            return false;
-        }
-        // Public struct functions should not expose private types.
-        if let Some(struct_visibility) = func_meta.type_id.and_then(|id| {
-            let struct_def = self.get_type(id);
-            let struct_def = struct_def.borrow();
-            self.find_struct_visibility(&struct_def)
-        }) {
-            return struct_visibility != ItemVisibility::Private;
-        }
-        // Standalone functions should be checked
-        true
-    }
-
-    /// Check that an item such as a struct field or type alias is not more visible than the type it refers to.
-    fn check_type_is_not_more_private_then_item(
-        &mut self,
-        name: &Ident,
-        visibility: ItemVisibility,
-        typ: &Type,
-        location: Location,
-    ) {
-        match typ {
-            Type::DataType(struct_type, generics) => {
-                let struct_type = struct_type.borrow();
-                let struct_module_id = struct_type.id.module_id();
-
-                // We only check this in types in the same crate. If it's in a different crate
-                // then it's either accessible (all good) or it's not, in which case a different
-                // error will happen somewhere else, but no need to error again here.
-                if struct_module_id.krate == self.crate_id {
-                    if let Some(aliased_visibility) = self.find_struct_visibility(&struct_type) {
-                        if aliased_visibility < visibility {
-                            self.push_err(ResolverError::TypeIsMorePrivateThenItem {
-                                typ: struct_type.name.to_string(),
-                                item: name.to_string(),
-                                location,
-                            });
-                        }
-                    }
-                }
-
-                for generic in generics {
-                    self.check_type_is_not_more_private_then_item(
-                        name, visibility, generic, location,
-                    );
-                }
-            }
-            Type::Tuple(types) => {
-                for typ in types {
-                    self.check_type_is_not_more_private_then_item(name, visibility, typ, location);
-                }
-            }
-            Type::Alias(alias_type, generics) => {
-                self.check_type_is_not_more_private_then_item(
-                    name,
-                    visibility,
-                    &alias_type.borrow().get_type(generics),
-                    location,
-                );
-            }
-            Type::CheckedCast { from, to } => {
-                self.check_type_is_not_more_private_then_item(name, visibility, from, location);
-                self.check_type_is_not_more_private_then_item(name, visibility, to, location);
-            }
-            Type::Function(args, return_type, env, _) => {
-                for arg in args {
-                    self.check_type_is_not_more_private_then_item(name, visibility, arg, location);
-                }
-                self.check_type_is_not_more_private_then_item(
-                    name,
-                    visibility,
-                    return_type,
-                    location,
-                );
-                self.check_type_is_not_more_private_then_item(name, visibility, env, location);
-            }
-            Type::Reference(typ, _) | Type::Array(_, typ) | Type::Slice(typ) => {
-                self.check_type_is_not_more_private_then_item(name, visibility, typ, location);
-            }
-            Type::InfixExpr(left, _op, right, _) => {
-                self.check_type_is_not_more_private_then_item(name, visibility, left, location);
-                self.check_type_is_not_more_private_then_item(name, visibility, right, location);
-            }
-            Type::FieldElement
-            | Type::Integer(..)
-            | Type::Bool
-            | Type::String(..)
-            | Type::FmtString(..)
-            | Type::Unit
-            | Type::Quoted(..)
-            | Type::TypeVariable(..)
-            | Type::Forall(..)
-            | Type::TraitAsType(..)
-            | Type::Constant(..)
-            | Type::NamedGeneric(..)
-            | Type::Error => (),
-        }
-    }
-
-    fn collect_struct_definitions(&mut self, structs: &BTreeMap<TypeId, UnresolvedStruct>) {
-        // This is necessary to avoid cloning the entire struct map
-        // when adding checks after each struct field is resolved.
-        let struct_ids = structs.keys().copied().collect::<Vec<_>>();
-
-        // Resolve each field in each struct.
-        // Each struct should already be present in the NodeInterner after def collection.
-        for (type_id, typ) in structs {
-            self.local_module = typ.module_id;
-
-            let fields = self.resolve_struct_fields(&typ.struct_def, *type_id);
-
-            if typ.struct_def.is_abi() {
-                for field in &fields {
-                    self.mark_type_as_used(&field.typ);
-                }
-            }
-
-            // Check that the a public struct doesn't have a private type as a public field.
-            if typ.struct_def.visibility != ItemVisibility::Private {
-                for field in &fields {
-                    let ident = Ident::from(Located::from(
-                        field.name.location(),
-                        format!("{}::{}", typ.struct_def.name, field.name),
-                    ));
-                    self.check_type_is_not_more_private_then_item(
-                        &ident,
-                        field.visibility,
-                        &field.typ,
-                        field.name.location(),
-                    );
-                }
-            }
-
-            if self.interner.is_in_lsp_mode() {
-                for (field_index, field) in fields.iter().enumerate() {
-                    let location = field.name.location();
-                    let reference_id = ReferenceId::StructMember(*type_id, field_index);
-                    self.interner.add_definition_location(reference_id, location);
-                }
-            }
-
-            self.interner.update_type(*type_id, |struct_def| {
-                struct_def.set_fields(fields);
-            });
-        }
-
-        // Check whether the struct fields have nested slices
-        // We need to check after all structs are resolved to
-        // make sure every struct's fields is accurately set.
-        for id in struct_ids {
-            let struct_type = self.interner.get_type(id);
-
-            // Only handle structs without generics as any generics args will be checked
-            // after monomorphization when performing SSA codegen
-            if struct_type.borrow().generics.is_empty() {
-                let fields = struct_type.borrow().get_fields(&[]).unwrap();
-                for (_, field_type, _) in fields.iter() {
-                    if field_type.is_nested_slice() {
-                        let location = struct_type.borrow().location;
-                        self.push_err(ResolverError::NestedSlices { location });
-                    }
-                }
-            }
-        }
-    }
-
-    pub fn resolve_struct_fields(
-        &mut self,
-        unresolved: &NoirStruct,
-        struct_id: TypeId,
-    ) -> Vec<StructField> {
-        self.recover_generics(|this| {
-            this.current_item = Some(DependencyId::Struct(struct_id));
-
-            this.resolving_ids.insert(struct_id);
-
-            let struct_def = this.interner.get_type(struct_id);
-            this.add_existing_generics(&unresolved.generics, &struct_def.borrow().generics);
-
-            let wildcard_allowed = false;
-            let fields = vecmap(&unresolved.fields, |field| {
-                let ident = &field.item.name;
-                let typ = &field.item.typ;
-                let visibility = field.item.visibility;
-                StructField {
-                    visibility,
-                    name: ident.clone(),
-                    typ: this.resolve_type(typ.clone(), wildcard_allowed),
-                }
-            });
-
-            this.resolving_ids.remove(&struct_id);
-
-            fields
-        })
-    }
-
-    fn collect_enum_definitions(&mut self, enums: &BTreeMap<TypeId, UnresolvedEnum>) {
-        for (type_id, typ) in enums {
-            self.local_module = typ.module_id;
-            self.generics.clear();
-
-            let datatype = self.interner.get_type(*type_id);
-            let datatype_ref = datatype.borrow();
-            let generics = datatype_ref.generic_types();
-            self.add_existing_generics(&typ.enum_def.generics, &datatype_ref.generics);
-
-            self.use_unstable_feature(UnstableFeature::Enums, datatype_ref.name.location());
-            drop(datatype_ref);
-
-            let self_type = Type::DataType(datatype.clone(), generics);
-            let self_type_id = self.interner.push_quoted_type(self_type.clone());
-            let location = typ.enum_def.location;
-            let unresolved =
-                UnresolvedType { typ: UnresolvedTypeData::Resolved(self_type_id), location };
-
-            datatype.borrow_mut().init_variants();
-            self.resolving_ids.insert(*type_id);
-
-            let wildcard_allowed = false;
-            for (i, variant) in typ.enum_def.variants.iter().enumerate() {
-                let parameters = variant.item.parameters.as_ref();
-                let types = parameters.map(|params| {
-                    vecmap(params, |typ| self.resolve_type(typ.clone(), wildcard_allowed))
-                });
-                let name = variant.item.name.clone();
-
-                let is_function = types.is_some();
-                let params = types.clone().unwrap_or_default();
-                datatype.borrow_mut().push_variant(EnumVariant::new(name, params, is_function));
-
-                self.define_enum_variant_constructor(
-                    &typ.enum_def,
-                    *type_id,
-                    &variant.item,
-                    types,
-                    i,
-                    &datatype,
-                    &self_type,
-                    unresolved.clone(),
-                );
-
-                let reference_id = ReferenceId::EnumVariant(*type_id, i);
-                let location = variant.item.name.location();
-                self.interner.add_definition_location(reference_id, location);
-            }
-
-            self.resolving_ids.remove(type_id);
-        }
-        self.generics.clear();
-    }
-
-    fn elaborate_global(&mut self, global: UnresolvedGlobal) {
-        let old_module = std::mem::replace(&mut self.local_module, global.module_id);
-        let old_item = self.current_item.take();
-
-        let global_id = global.global_id;
-        self.current_item = Some(DependencyId::Global(global_id));
-        let let_stmt = global.stmt_def;
-
-        let name = if self.interner.is_in_lsp_mode() {
-            Some(let_stmt.pattern.name_ident().to_string())
-        } else {
-            None
-        };
-
-        let location = let_stmt.pattern.location();
-
-        if !self.in_contract() {
-            for attr in &let_stmt.attributes {
-                if matches!(attr.kind, SecondaryAttributeKind::Abi(_)) {
-                    self.push_err(ResolverError::AbiAttributeOutsideContract {
-                        location: attr.location,
-                    });
-                }
-            }
-        }
-
-        if !let_stmt.comptime && matches!(let_stmt.pattern, Pattern::Mutable(..)) {
-            self.push_err(ResolverError::MutableGlobal { location });
-        }
-
-        let (let_statement, _typ) = self
-            .elaborate_in_comptime_context(|this| this.elaborate_let(let_stmt, Some(global_id)));
-
-        if let_statement.r#type.contains_reference() {
-            self.push_err(ResolverError::ReferencesNotAllowedInGlobals { location });
-        }
-
-        let let_statement = HirStatement::Let(let_statement);
-
-        let statement_id = self.interner.get_global(global_id).let_statement;
-        self.interner.replace_statement(statement_id, let_statement);
-
-        self.elaborate_comptime_global(global_id);
-
-        if let Some(name) = name {
-            self.interner.register_global(global_id, name, location, global.visibility);
-        }
-
-        self.local_module = old_module;
-        self.current_item = old_item;
-    }
-
-    fn elaborate_comptime_global(&mut self, global_id: GlobalId) {
-        let let_statement = self
-            .interner
-            .get_global_let_statement(global_id)
-            .expect("Let statement of global should be set by elaborate_global_let");
-
-        let global = self.interner.get_global(global_id);
-        let definition_id = global.definition_id;
-        let location = global.location;
-        let mut interpreter = self.setup_interpreter();
-
-        if let Err(error) = interpreter.evaluate_let(let_statement) {
-            let error: CompilationError = error.into();
-            self.push_err(error);
-        } else {
-            let value = interpreter
-                .lookup_id(definition_id, location)
-                .expect("The global should be defined since evaluate_let did not error");
-
-            self.debug_comptime(location, |interner| value.display(interner).to_string());
-
-            self.interner.get_global_mut(global_id).value = GlobalValue::Resolved(value);
-        }
-    }
-
-    /// If the given global is unresolved, elaborate it and return true
-    fn elaborate_global_if_unresolved(&mut self, global_id: &GlobalId) -> bool {
-        if let Some(global) = self.unresolved_globals.remove(global_id) {
-            self.elaborate_global(global);
-            true
-        } else {
-            false
-        }
     }
 
     fn define_function_metas(
@@ -2519,58 +1268,6 @@ impl<'context> Elaborator<'context> {
             })
     }
 
-    /// Check that all the generics show up in any of `types` (if they don't, we produce an error),
-    /// or in any of a where clause associated type binding.
-    fn check_generics_appear_in_types(
-        &mut self,
-        generics: &[UnresolvedGeneric],
-        types: &[&UnresolvedType],
-        where_clause: &[UnresolvedTraitConstraint],
-    ) {
-        if generics.is_empty() {
-            return;
-        }
-
-        // Turn each generic into an Ident
-        let mut idents = HashSet::new();
-        for generic in generics {
-            match generic {
-                UnresolvedGeneric::Variable(ident, _)
-                | UnresolvedGeneric::Numeric { ident, typ: _ } => match ident {
-                    IdentOrQuotedType::Ident(ident) => {
-                        idents.insert(ident.clone());
-                    }
-                    IdentOrQuotedType::Quoted(quoted_type_id, location) => {
-                        if let Type::NamedGeneric(NamedGeneric { name, .. }) =
-                            self.interner.get_quoted_type(*quoted_type_id).follow_bindings()
-                        {
-                            idents.insert(Ident::new(name.to_string(), *location));
-                        }
-                    }
-                },
-            }
-        }
-
-        // Remove the ones that show up in `self_type`
-        let mut visitor =
-            RemoveGenericsAppearingInTypeVisitor { interner: self.interner, idents: &mut idents };
-        for typ in types {
-            typ.accept(&mut visitor);
-        }
-
-        // Removes the ones that show up in associated type bindings in the where clause
-        for where_clause in where_clause {
-            for (_name, typ) in &where_clause.trait_bound.trait_generics.named_args {
-                typ.accept(&mut visitor);
-            }
-        }
-
-        // The ones that remain are not mentioned in the impl: it's an error.
-        for ident in idents {
-            self.push_err(ResolverError::UnconstrainedTypeParameter { ident });
-        }
-    }
-
     /// Register a use of the given unstable feature. Errors if the feature has not
     /// been explicitly enabled in this package.
     pub fn use_unstable_feature(&mut self, feature: UnstableFeature, location: Location) {
@@ -2604,87 +1301,6 @@ impl<'context> Elaborator<'context> {
         let ret = f(self);
         let errored = self.errors.iter().skip(previous_errors).any(|error| error.is_error());
         (errored, ret)
-    }
-}
-
-struct RemoveGenericsAppearingInTypeVisitor<'interner, 'ident> {
-    interner: &'interner NodeInterner,
-    idents: &'ident mut HashSet<Ident>,
-}
-
-impl RemoveGenericsAppearingInTypeVisitor<'_, '_> {
-    fn visit_type(&mut self, typ: &Type) {
-        match typ {
-            Type::Array(length, element) => {
-                self.visit_type(length);
-                self.visit_type(element);
-            }
-            Type::Slice(element) => {
-                self.visit_type(element);
-            }
-            Type::FmtString(length, element) => {
-                self.visit_type(length);
-                self.visit_type(element);
-            }
-            Type::Tuple(items) | Type::DataType(_, items) | Type::Alias(_, items) => {
-                for item in items {
-                    self.visit_type(item);
-                }
-            }
-            Type::TraitAsType(_, _, trait_generics) => {
-                for generic in &trait_generics.ordered {
-                    self.visit_type(generic);
-                }
-                for named_type in &trait_generics.named {
-                    self.visit_type(&named_type.typ);
-                }
-            }
-            Type::NamedGeneric(named_generic) => {
-                let ident = Ident::new(named_generic.name.to_string(), Location::dummy());
-                self.idents.remove(&ident);
-            }
-            Type::CheckedCast { from, to } => {
-                self.visit_type(from);
-                self.visit_type(to);
-            }
-            Type::Function(args, ret, env, _) => {
-                for arg in args {
-                    self.visit_type(arg);
-                }
-                self.visit_type(ret);
-                self.visit_type(env);
-            }
-            Type::Reference(typ, _) => {
-                self.visit_type(typ);
-            }
-            Type::InfixExpr(lhs, _, rhs, _) => {
-                self.visit_type(lhs);
-                self.visit_type(rhs);
-            }
-            Type::Unit
-            | Type::Bool
-            | Type::Integer(..)
-            | Type::FieldElement
-            | Type::String(_)
-            | Type::Constant(..)
-            | Type::Quoted(_)
-            | Type::Forall(..)
-            | Type::TypeVariable(_)
-            | Type::Error => (),
-        }
-    }
-}
-
-impl Visitor for RemoveGenericsAppearingInTypeVisitor<'_, '_> {
-    fn visit_path(&mut self, path: &Path) {
-        if let Some(ident) = path.as_ident() {
-            self.idents.remove(ident);
-        }
-    }
-
-    fn visit_resolved_type(&mut self, quoted_type_id: QuotedTypeId, _: Location) {
-        let typ = self.interner.get_quoted_type(quoted_type_id);
-        self.visit_type(typ);
     }
 }
