+++ compiler/noirc_frontend/src/elaborator/trait_impls.rs
@@ -1,11 +1,16 @@
 use crate::{
-    Kind, NamedGeneric, ResolvedGeneric,
+    Kind, NamedGeneric, ResolvedGeneric, Shared, TypeBindings, TypeVariable,
     ast::{Ident, UnresolvedType, UnresolvedTypeData, UnresolvedTypeExpression},
+    elaborator::types::bind_ordered_generics,
     graph::CrateId,
-    hir::def_collector::{
-        dc_crate::{CompilationError, UnresolvedTraitImpl},
-        errors::DefCollectorErrorKind,
+    hir::{
+        def_collector::{
+            dc_crate::{CompilationError, UnresolvedTraitImpl},
+            errors::DefCollectorErrorKind,
+        },
+        resolution::errors::ResolverError,
     },
+    hir_def::traits::{NamedType, ResolvedTraitBound, TraitImpl},
     node_interner::TraitImplId,
 };
 use crate::{
@@ -15,11 +20,181 @@ use crate::{
     node_interner::{FuncId, TraitId},
 };
 
+use iter_extended::vecmap;
+use rustc_hash::FxHashMap as HashMap;
 use rustc_hash::FxHashSet as HashSet;
 
 use super::Elaborator;
 
 impl Elaborator<'_> {
+    pub(super) fn collect_trait_impl(&mut self, trait_impl: &mut UnresolvedTraitImpl) {
+        self.local_module = trait_impl.module_id;
+        self.current_trait_impl = trait_impl.impl_id;
+
+        let self_type = trait_impl.methods.self_type.clone();
+        let self_type =
+            self_type.expect("Expected struct type to be set before collect_trait_impl");
+
+        self.self_type = Some(self_type.clone());
+        let self_type_location = trait_impl.object_type.location;
+
+        if matches!(self_type, Type::Reference(..)) {
+            self.push_err(DefCollectorErrorKind::ReferenceInTraitImpl {
+                location: self_type_location,
+            });
+        }
+
+        self.check_generics_appear_in_types(
+            &trait_impl.generics,
+            &[&trait_impl.r#trait, &trait_impl.object_type],
+            &trait_impl.where_clause,
+        );
+
+        if let Some(trait_id) = trait_impl.trait_id {
+            self.generics = trait_impl.resolved_generics.clone();
+
+            let where_clause = self.resolve_trait_constraints(&trait_impl.where_clause);
+
+            // Now solve the actual type of associated types
+            // (before this we only declared them without knowing their type)
+            if let Some(trait_impl_id) = trait_impl.impl_id {
+                let unresolved_associated_types =
+                    std::mem::take(&mut trait_impl.unresolved_associated_types);
+                let mut unresolved_associated_types =
+                    unresolved_associated_types.into_iter().collect::<HashMap<_, _>>();
+
+                let associated_types =
+                    self.interner.get_associated_types_for_impl(trait_impl_id).to_vec();
+                for associated_type in &associated_types {
+                    let Type::NamedGeneric(named_generic) = &associated_type.typ else {
+                        // This can happen if the associated type is specified directly in the impl trait generics,
+                        // This can't be done in code, but it could happen with unquoted types.
+                        continue;
+                    };
+
+                    let Some(unresolved_type) =
+                        unresolved_associated_types.remove(&associated_type.name)
+                    else {
+                        // This too can happen if the associated type is specified directly in the impl trait generics,
+                        // like `impl<H> BuildHasher<H = H>`, where `H` is a named generic but its resolution isn't delayed.
+                        // This can't be done in code, but it could happen with unquoted types.
+                        continue;
+                    };
+                    let wildcard_allowed = false;
+                    let resolved_type = self.resolve_type_with_kind(
+                        unresolved_type,
+                        &associated_type.typ.kind(),
+                        wildcard_allowed,
+                    );
+                    named_generic.type_var.bind(resolved_type);
+                }
+            }
+
+            let trait_ = self.interner.get_trait(trait_id);
+
+            // If there are bounds on the trait's associated types, check them now
+            let associated_type_bounds = &trait_.associated_type_bounds;
+            if !associated_type_bounds.is_empty() {
+                let associated_type_bounds = associated_type_bounds.clone();
+                let named_generics = self
+                    .interner
+                    .get_associated_types_for_impl(trait_impl.impl_id.unwrap())
+                    .to_vec();
+                for named_generic in named_generics {
+                    let Some(bounds) = associated_type_bounds.get(named_generic.name.as_str())
+                    else {
+                        continue;
+                    };
+                    let object_type = &named_generic.typ;
+                    for bound in bounds {
+                        if let Err(error) = self.interner.lookup_trait_implementation(
+                            object_type,
+                            bound.trait_id,
+                            &bound.trait_generics.ordered,
+                            &bound.trait_generics.named,
+                        ) {
+                            self.push_trait_constraint_error(
+                                object_type,
+                                error,
+                                named_generic.name.location(),
+                            );
+                        }
+                    }
+                }
+            }
+
+            self.remove_trait_constraints_from_scope(where_clause.iter());
+
+            self.collect_trait_impl_methods(trait_id, trait_impl, &where_clause);
+
+            let location = trait_impl.object_type.location;
+            self.declare_methods_on_struct(Some(trait_id), &mut trait_impl.methods, location);
+
+            let trait_visibility = self.interner.get_trait(trait_id).visibility;
+
+            let methods = trait_impl.methods.function_ids();
+            for func_id in &methods {
+                self.interner.set_function_trait(*func_id, self_type.clone(), trait_id);
+
+                // A trait impl method has the same visibility as its trait
+                let modifiers = self.interner.function_modifiers_mut(func_id);
+                modifiers.visibility = trait_visibility;
+            }
+
+            let trait_generics = trait_impl.resolved_trait_generics.clone();
+            let ident = match &trait_impl.r#trait.typ {
+                UnresolvedTypeData::Named(trait_path, _, _) => trait_path.last_ident(),
+                UnresolvedTypeData::Resolved(quoted_type_id) => {
+                    let typ = self.interner.get_quoted_type(*quoted_type_id);
+                    let name = if let Type::TraitAsType(_, name, _) = typ {
+                        name.to_string()
+                    } else {
+                        typ.to_string()
+                    };
+                    Ident::new(name, trait_impl.r#trait.location)
+                }
+                _ => {
+                    // We don't error in this case because an error will be produced later on when
+                    // solving the trait impl trait type
+                    Ident::new(trait_impl.r#trait.to_string(), trait_impl.r#trait.location)
+                }
+            };
+
+            let resolved_trait_impl = Shared::new(TraitImpl {
+                ident,
+                location,
+                typ: self_type.clone(),
+                trait_id,
+                trait_generics,
+                file: trait_impl.file_id,
+                crate_id: self.crate_id,
+                where_clause,
+                methods,
+            });
+
+            let generics = vecmap(&self.generics, |generic| generic.type_var.clone());
+
+            if let Err(prev_location) = self.interner.add_trait_implementation(
+                self_type.clone(),
+                trait_id,
+                trait_impl.impl_id.expect("impl_id should be set in define_function_metas"),
+                generics,
+                resolved_trait_impl,
+            ) {
+                self.push_err(DefCollectorErrorKind::OverlappingImpl {
+                    typ: self_type.clone(),
+                    location: self_type_location,
+                    prev_location,
+                });
+            }
+        }
+
+        self.generics.clear();
+
+        self.current_trait_impl = None;
+        self.self_type = None;
+    }
+
     pub(super) fn collect_trait_impl_methods(
         &mut self,
         trait_id: TraitId,
@@ -265,5 +440,213 @@ impl Elaborator<'_> {
             associated_types.push((name, typ, Kind::Any));
         }
         associated_types
+    }
+
+    pub(super) fn add_trait_impl_assumed_trait_implementations(
+        &mut self,
+        impl_id: Option<TraitImplId>,
+    ) {
+        if let Some(impl_id) = impl_id {
+            if let Some(trait_implementation) = self.interner.try_get_trait_implementation(impl_id)
+            {
+                for trait_constrain in &trait_implementation.borrow().where_clause {
+                    let trait_bound = &trait_constrain.trait_bound;
+                    self.add_trait_bound_to_scope(
+                        trait_bound.location,
+                        &trait_constrain.typ,
+                        trait_bound,
+                        trait_bound.trait_id,
+                    );
+                }
+            }
+        }
+    }
+
+    pub(super) fn remove_trait_impl_assumed_trait_implementations(
+        &mut self,
+        impl_id: Option<TraitImplId>,
+    ) {
+        if let Some(impl_id) = impl_id {
+            if let Some(trait_implementation) = self.interner.try_get_trait_implementation(impl_id)
+            {
+                for trait_constrain in &trait_implementation.borrow().where_clause {
+                    self.interner.remove_assumed_trait_implementations_for_trait(
+                        trait_constrain.trait_bound.trait_id,
+                    );
+                }
+            }
+        }
+    }
+
+    pub(super) fn check_trait_impl_where_clause_matches_trait_where_clause(
+        &mut self,
+        trait_impl: &UnresolvedTraitImpl,
+    ) {
+        let Some(trait_id) = trait_impl.trait_id else {
+            return;
+        };
+
+        let Some(the_trait) = self.interner.try_get_trait(trait_id) else {
+            return;
+        };
+
+        if the_trait.where_clause.is_empty() {
+            return;
+        }
+
+        let impl_trait = the_trait.name.to_string();
+
+        let mut bindings = TypeBindings::default();
+        bind_ordered_generics(
+            &the_trait.generics,
+            &trait_impl.resolved_trait_generics,
+            &mut bindings,
+        );
+
+        // Check that each of the trait's where clause constraints is satisfied
+        for trait_constraint in the_trait.where_clause.clone() {
+            let Some(trait_constraint_trait) =
+                self.interner.try_get_trait(trait_constraint.trait_bound.trait_id)
+            else {
+                continue;
+            };
+            let trait_constraint_trait_name = trait_constraint_trait.name.to_string();
+
+            let mut trait_constraint = trait_constraint.clone();
+            trait_constraint.apply_bindings(&bindings);
+
+            let trait_constraint_type = trait_constraint.typ;
+            let trait_bound = trait_constraint.trait_bound;
+
+            let mut named_generics = trait_bound.trait_generics.named.clone();
+
+            // If the trait bound is over a trait that has associated types, the ones that
+            // aren't explicit will be in `named_generics` as implicitly added ones.
+            // If they are unbound, they won't be bound until monomorphization, in which cae
+            // the below trait implementation lookup will fail (an unbound named generic will
+            // never unify in this case). In this case we replace them with fresh type variables
+            // so they'll unify (the bindings aren't applied here so this is fine).
+            // If they are bound though, we won't replace them as we want to ensure the binding
+            // matches.
+            //
+            // `bindings` is passed here because these implicitly added named generics might
+            // have a constraint on them later on and we want to remember what type they ended
+            // up being.
+            self.replace_implicitly_added_unbound_named_generics_with_fresh_type_variables(
+                &mut named_generics,
+                &mut bindings,
+            );
+
+            match self.interner.try_lookup_trait_implementation(
+                &trait_constraint_type,
+                trait_bound.trait_id,
+                &trait_bound.trait_generics.ordered,
+                &named_generics,
+            ) {
+                Ok((_, impl_bindings, impl_instantiation_bindings)) => {
+                    bindings.extend(impl_bindings);
+                    bindings.extend(impl_instantiation_bindings);
+                }
+                Err(_) => {
+                    let missing_trait =
+                        format!("{}{}", trait_constraint_trait_name, trait_bound.trait_generics);
+                    self.push_err(ResolverError::TraitNotImplemented {
+                        impl_trait: impl_trait.clone(),
+                        missing_trait,
+                        type_missing_trait: trait_constraint_type.to_string(),
+                        location: trait_impl.object_type.location,
+                        missing_trait_location: trait_bound.location,
+                    });
+                }
+            }
+        }
+    }
+
+    fn replace_implicitly_added_unbound_named_generics_with_fresh_type_variables(
+        &mut self,
+        named_generics: &mut [NamedType],
+        bindings: &mut TypeBindings,
+    ) {
+        for named_type in named_generics.iter_mut() {
+            match &named_type.typ {
+                Type::NamedGeneric(NamedGeneric { type_var, implicit: true, .. })
+                    if type_var.borrow().is_unbound() =>
+                {
+                    let type_var_id = type_var.id();
+                    let new_type_var_id = self.interner.next_type_variable_id();
+                    let kind = type_var.kind();
+                    let new_type_var = TypeVariable::unbound(new_type_var_id, kind.clone());
+                    named_type.typ = Type::TypeVariable(new_type_var.clone());
+                    bindings.insert(type_var_id, (new_type_var, kind, named_type.typ.clone()));
+                }
+                _ => (),
+            };
+        }
+    }
+
+    pub(super) fn check_parent_traits_are_implemented(&mut self, trait_impl: &UnresolvedTraitImpl) {
+        let Some(trait_id) = trait_impl.trait_id else {
+            return;
+        };
+
+        let Some(object_type) = &trait_impl.resolved_object_type else {
+            return;
+        };
+
+        let Some(the_trait) = self.interner.try_get_trait(trait_id) else {
+            return;
+        };
+
+        if the_trait.trait_bounds.is_empty() {
+            return;
+        }
+
+        let impl_trait = the_trait.name.to_string();
+
+        let mut bindings = TypeBindings::default();
+        bind_ordered_generics(
+            &the_trait.generics,
+            &trait_impl.resolved_trait_generics,
+            &mut bindings,
+        );
+
+        // Note: we only check if the immediate parents are implemented, we don't check recursively.
+        // Why? If a parent isn't implemented, we get an error. If a parent is implemented, we'll
+        // do the same check for the parent, so this trait's parents parents will be checked, so the
+        // recursion is guaranteed.
+        for parent_trait_bound in the_trait.trait_bounds.clone() {
+            let Some(parent_trait) = self.interner.try_get_trait(parent_trait_bound.trait_id)
+            else {
+                continue;
+            };
+
+            let parent_trait_bound = ResolvedTraitBound {
+                trait_generics: parent_trait_bound
+                    .trait_generics
+                    .map(|typ| typ.substitute(&bindings)),
+                ..parent_trait_bound
+            };
+
+            if self
+                .interner
+                .try_lookup_trait_implementation(
+                    object_type,
+                    parent_trait_bound.trait_id,
+                    &parent_trait_bound.trait_generics.ordered,
+                    &parent_trait_bound.trait_generics.named,
+                )
+                .is_err()
+            {
+                let missing_trait =
+                    format!("{}{}", parent_trait.name, parent_trait_bound.trait_generics);
+                self.push_err(ResolverError::TraitNotImplemented {
+                    impl_trait: impl_trait.clone(),
+                    missing_trait,
+                    type_missing_trait: trait_impl.object_type.to_string(),
+                    location: trait_impl.object_type.location,
+                    missing_trait_location: parent_trait_bound.location,
+                });
+            }
+        }
     }
 }
